fix: move Comments on Issues/PRs from lerna publish to version
