+++ packages/version/src/__mocks__/get-commits-since-last-release.ts
@@ -1,4 +1,5 @@
 import { vi } from 'vitest';
 
 export const getCommitsSinceLastRelease = vi.fn(() => Promise.resolve());
+export const getLastTagCommit = vi.fn();
 export const getLastTagDetails = vi.fn(() => ({ tagHash: undefined, tagDate: undefined, tagRefCount: 0 }));
