+++ packages/version/src/constant.ts
