+++ packages/version/src/conventional-commits/__tests__/conventional-commits.spec.ts
@@ -542,9 +542,8 @@ describe('conventional-commits', () => {
         updateChangelog(rootPkg as Package, 'root', { version: '1.0.1' }).then(getFileContent),
       ]);
 
-      // not sure why CI sometimes replaces `*` with `-`
-      expect(leafChangelogContent.replace('- I should', '* I should')).toMatchSnapshot('leaf');
-      expect(rootChangelogContent.replace('- I should', '* I should')).toMatchSnapshot('root');
+      expect(leafChangelogContent).toMatchSnapshot('leaf');
+      expect(rootChangelogContent).toMatchSnapshot('root');
     });
 
     it('supports custom tagPrefix in fixed mode', async () => {
